requirements.txt
altair==5.4.1
attrs==24.2.0
babel==2.16.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.18.0
et-xmlfile==1.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.8.1
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdfkit==1.0.0
pillow==10.4.0
protobuf==5.28.1
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pypandoc==1.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pywin32==306
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==4.0.2
